+++ spec/dummy/db/migrate/20160415194001_create_posts.rb
@@ -1,13 +1,24 @@
 # frozen_string_literal: true
 
 class CreatePosts < ActiveRecord::Migration[5.0]
+  include DatabaseHelpers
+
   def change
     create_table :posts do |t|
       t.string :title
       t.integer :rating
       t.boolean :active
-      t.json :meta
-      t.json :data
+
+      if postgresql?
+        t.jsonb :meta
+        t.jsonb :data
+      end
+
+      if mysql?
+        t.json :meta
+        t.json :data
+      end
+
       t.references :user
 
       t.timestamps null: false
