+++ spec/dummy/db/migrate/20170220132233_create_articles.rb
@@ -1,13 +1,22 @@
 # frozen_string_literal: true
 
 class CreateArticles < ActiveRecord::Migration[5.0]
+  include DatabaseHelpers
+
   def change
     create_table :articles do |t|
       t.string :title
       t.integer :rating
       t.boolean :active
       t.references :user, foreign_key: true
-      t.jsonb :log_data
+
+      if postgresql?
+        t.jsonb :log_data
+      end
+
+      if mysql?
+        t.json :log_data
+      end
 
       t.timestamps null: false
     end
