+++ spec/dummy/db/migrate/20170220153105_create_comments.rb
@@ -1,10 +1,20 @@
 # frozen_string_literal: true
 
 class CreateComments < ActiveRecord::Migration[5.0]
+  include DatabaseHelpers
+
   def change
     create_table :comments do |t|
       t.text :content
-      t.jsonb :log_data
+
+      if postgresql?
+        t.jsonb :log_data
+      end
+
+      if mysql?
+        t.json :log_data
+      end
+
       t.references :article, foreign_key: true
 
       t.timestamps null: false
