--- spec/dummy/db/mysql/20170220153105_create_comments.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-class CreateComments < ActiveRecord::Migration[5.0]
-  def change
-    create_table :comments do |t|
-      t.text :content
-      t.json :log_data
-      t.references :article, foreign_key: true
-
-      t.timestamps null: false
-    end
-  end
-end
